moved requestCaller into util package
